GraviPet.jl

GraviPet is the General Relativistic Astrophysics Visualization, Initialization, and Postprocessing Efficient Toolkit.

Domains

GraviPet.Domains.Domain — Type

abstract type Domain{T}

A computational domain of type T. The domain of a function is the type of its argument(s). For example, Domain{Float64} would describe a one-dimensional domain, and Domain{SVector{3,Float64}} would describe a three-dimensional domain.

This type can also describe a codomain, i.e. the result type of a function. A codomain of Domain{Float64} describes a scalar function, and a codomain of Domain{SVector{3,Float64}} describes a vector-valued function.

See also Interval and Box for concrete domain types.

GraviPet.Domains.expanded — Method

expanded(dom::Domain, delta)::Domain
expanded(dom::Domain, delta_lo, delta_hi)::Domain

Expand the domain dom. The lower and upper domain bounds are moved by either delta, or delta_lo and delta_hi, respectively. Positive deltas enlarge the domain, negative values shrink it. Domains must be non-empty.

expanded(dom, delta) is equivalent to expanded(dom, delta, delta).

See also shifted.

GraviPet.Domains.shifted — Method

shifted(dom::Domain, shift)::Domain

Shift the domain dom by moving both domain bounds in the same direction. This does not change the size of the domain.

shfited(dom, shift) is equivalent to expanded(dom, -shift, +shift).

See also expanded.

Intervals

GraviPet.Intervals.Interval — Type

struct Interval{T} <: Domain{T}

A one-dimensional Domain, i.e. an interval. T should be a real-valued scalar type such as Float64.

GraviPet.Intervals.Interval — Method

Interval{T}(first::Real, last::Real)
Interval(first::Real, last::Real)

Create an interval with the specified bounds. The bounds can be infinity. The constraint first < last must hold.

GraviPet.Intervals.Interval — Method

Interval{T}()

Create a zero-valued interval, i.e. an interval where both lower and upper bounds are zero. This is also a convenient way to provide a "skeleton" for a domain without having to specify any bounds.

Boxes

GraviPet.Boxes.Box — Type

struct Box{D,T} <: Domain{SVector{D,T}}

A multi-dimensional Domains.Domain, i.e. a box or hyper-rectangle. The number of dimensions D must be a non-negative integer. T should be a real-valued scalar type such as Float64. The domain's element type is SVector{D,T}.

Zero-dimensional and one-dimensional domains are supported. Zero-dimensional domains contain only a single point (the origin). One-dimensional domains correspond to intervals (see Interval).

GraviPet.Boxes.Box — Method

Box{D,T}(first::SVector{D}, last::SVector{D})
Box(first::SVector{D}, last::SVector{D})

Create a box domain with the specified lower and upper bounds. The bounds are given as vectors, and (some of) their elements can be infinity. The constraint first < last must hold in each dimension.

GraviPet.Boxes.Box — Method

Box{D,T}()

Create a zero-valued box domain, i.e. a box where both lower and upper bounds are zero. This is also a convenient way to provide a "skeleton" for a domain without having to specify any bounds.

Categories

GraviPet.Categories.Category — Method

(cat::Category)(x)::eltype(codomain(cat))

Evaluate the function cat at point x. This is written as a function call cat(x). See also evaluate.

GraviPet.Categories.evaluate — Method

evaluate(cat::Category, x)::eltype(codomain(cat))

Evaluate the function cat at point x. This can also be written as a function call cat(x).

GraviPet.Categories.integrate — Method

integrate(cat::Category)::eltype(codomain(cat))

Integrate the function cat over its domain.

GraviPet.Categories.project — Method

project(dst::T, src::Category)::T where {T<:Category}

Project the function src onto the representation given by dst. This can be used e.g. to sample Julia functions on a given grid, or to resample functiont on a different grid. dst can be a skeleton representation.

Julia Functions

GraviPet.JuliaFunctions.JuliaFunction — Method

JuliaFunction{DS,S,DT,T}(name::AbstractString, domain::Box{DS,S}, codomain::Box{DT,T})
JuliaFunction(name::AbstractString, domain::Box{DS,S}, codomain::Box{DT,T})

Create a zero-valued Julia function, i.e. a Julia function that is zero everywhere. The codomain can be a zero-valued domain created via Box{DT,T}().

1D Grid Functions

GraviPet.GridFunctions1D.GridFunction1D — Method

GridFunction1D{S,T}(name::AbstractString, domain::Interval{S}, codomain::Interval{T}, npoints::Int)
GridFunction1D(name::AbstractString, domain::Interval{S}, codomain::Interval{T}, npoints::Int)

Create a zero-valued grid function, i.e. a grid function that is zero everywhere. The codomain can be a zero-valued domain created via Interval{T}().

npoints specifies the resolution of the grid function. This makes it convenient to use zero-valued grid functions as "templates" or "skeletons" when creating other grid functions.

This zero-valued grid function is represented efficiently and does not store one element per grid point.

Multi-dimensional Grid Functions

GraviPet.GridFunctions.GridFunction — Method

GridFunction{DS,S,DT,T}(name::AbstractString, domain::Box{DS,S}, codomain::Box{DT,T}, grid_size::SVector{DS,Int})
GridFunction(name::AbstractString, domain::Box{DS,S}, codomain::Box{DT,T}, grid_size::SVector{DS,Int})

Create a zero-valued grid function, i.e. a grid function that is zero everywhere. The codomain can be a zero-valued domain created via Box{DT,T}().

grid_size specifies the resolution of the grid function. This makes it convenient to use zero-valued grid functions as "templates" or "skeletons" when creating other grid functions.

This zero-valued grid function is represented efficiently and does not store one element per grid point.

Blocked Functions (Domain Decomposition)

Threaded Functions

Distributed Functions

Miscellaneous

Base.:== — Method

==(dom1::Domain, dom2::Domain)::Bool

Equality for domains. Domains are equal if the have the same number of dimensions (see ndims), and the same lower and upper bounds (see first and last). The element types might be different if they can be compared for equality, e.g. Float64 and BigFloat.

Incompatible domains (living in different number of dimensions) are treated as unequal.

Base.last — Method

last(dom::Domain{T})::T

Return the upper bound of the domain. (At the moment domains are assumed to be cuboid, i.e. to form a box or hyper-rectangle.) Domain bounds can be infinity.

See also first.

Base.isdisjoint — Method

isdisjoint(dom1::Domain, dom2::Domain)::Bool

Test whether dom1 is disjoint from dom2. Only compatible domains (which live in the same number of dimensions) can be compared.

Base.eltype — Method

eltype(dom::Domain)::Type

Return the element type of a Domain.

Base.ndims — Method

ndims(dom::Domain)::Int

Return the number of dimensions of a Domain. Scalar domains are one-dimensional. Zero-dimensional domains consist of a single point only.

Base.issubset — Method

issubset(dom1::Domain, dom2::Domain)::Bool

Test whether dom1 is a subset of dom2. Only compatible domains (which live in the same number of dimensions) can be compared.

Base.in — Method

in(elem, dom::Domain)::Bool

Test whether elem is inside domain dom.

Base.first — Method

first(dom::Domain{T})::T

Return the lower bound of the domain. (At the moment domains are assumed to be cuboid, i.e. to form a box or hyper-rectangle.) Domain bounds can be infinity.

See also last.

Public Helpers

GraviPet.Funs.Fun — Type

struct Fun{Tuple{Ts...},R}

A function that takes inputs with the types Ts and returns a result of type R.